move donation button to new file
