+++ client/src/App.js
@@ -10,7 +10,6 @@ import {
 } from './components';
 import store from './redux/app/store';
 import { getUserData } from './redux/feature/user/userSlice';
-import Campaign from './components/common/campaignCard';
 
 function App() {
   useEffect(() => {
