+++ client/src/components/index.jsx
@@ -3,7 +3,7 @@ import Nav from './Nav';
 import Signin from './signin';
 import SignUp from './signUp';
 import Review from './Review';
-import DonationButton from './donateForm';
+import DonationButton from './donateForm/DonateButton';
 import ReportsForm from './ReportsForm';
 
 export {
